Organizing the directory of the files
