+++ frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { AppProvider } from "./hooks/useAppContext";
+import { AppProvider } from "@/hooks/useAppContext";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
