+++ frontend/src/components/Icons/LikeIcon.tsx
