+++ frontend/src/components/Icons/ViewIcon.tsx
