+++ frontend/src/components/Sugestions/RhsSugestions/index.tsx
