+++ frontend/src/components/User/SignInForm/index.tsx
@@ -1,8 +1,7 @@
 "use cliente";
 
-import { useAppContext } from "@/app/hooks/useAppContext";
-import { User } from "@/app/types";
-import { postUser } from "@/app/utils/apiCalls";
+import { useAppContext } from "@/hooks/useAppContext";
+import { signIn } from "@/utils/apiCalls";
 import Image from "next/image";
 import React, { useCallback, useState } from "react";
 import xIcon from "../../../../public/x-logo.png";
@@ -13,20 +12,15 @@ type Props = {
   handleClose: (value: boolean) => void;
 };
 
-const UserForm: React.FC<Props> = ({ handleClose }) => {
+const SignInForm: React.FC<Props> = ({ handleClose }) => {
   const { setIsLoading } = useAppContext();
-  const [user, setUser] = useState<User>({
-    email: "",
-    userName: "",
-    birthDate: "",
-    passwordHash: "",
-  });
+  const [user, setUser] = useState({ userName: "", passwordHash: "" });
 
   const onSubmit = useCallback(async () => {
-    handleClose(false);
     setIsLoading(true);
-    await postUser(user);
+    await signIn(user);
     setIsLoading(false);
+    handleClose(false);
   }, [handleClose, setIsLoading, user]);
 
   return (
@@ -34,23 +28,16 @@ const UserForm: React.FC<Props> = ({ handleClose }) => {
       <form>
         <div className="flex flex-col items-center gap-3 text-black">
           <Image height={30} width={30} alt="x image" src={xIcon} />
-          <h3>Create your account</h3>
+          <h3>Sign in to X</h3>
         </div>
         <div className="flex flex-col gap-2 px-4 py-4">
           <Input
             label="Name"
             form={true}
             value={user.userName}
-            onChange={(value) =>
-              setUser((prev) => ({ ...prev, userName: value }))
+            onChange={(values) =>
+              setUser((prev) => ({ ...prev, userName: values }))
             }
-          />
-          <Input
-            label="Email"
-            type="email"
-            form={true}
-            value={user.email}
-            onChange={(value) => setUser((prev) => ({ ...prev, email: value }))}
           />
           <Input
             label="Password"
@@ -61,20 +48,11 @@ const UserForm: React.FC<Props> = ({ handleClose }) => {
               setUser((prev) => ({ ...prev, passwordHash: value }))
             }
           />
-          <Input
-            label="Date of birth"
-            type="date"
-            form={true}
-            value={user.birthDate}
-            onChange={(value) =>
-              setUser((prev) => ({ ...prev, birthDate: value }))
-            }
-          />
         </div>
         <Button
           className="py-2 px-4 m-4 "
           backgroundColor="secondary"
-          label="Submit"
+          label="Sign in"
           onClick={onSubmit}
           applyMinWidth
         />
@@ -83,4 +61,4 @@ const UserForm: React.FC<Props> = ({ handleClose }) => {
   );
 };
 
-export default UserForm;
+export default SignInForm;
