+++ frontend/src/components/types.ts
