+++ frontend/src/utils/apiCalls.ts
@@ -1,4 +1,4 @@
-import { Tweet, Comment, User } from "../types";
+import { Comment, Tweet, User } from "@/types";
 
 const url = `${process.env.NEXT_PUBLIC_API_URL ?? "http://localhost:4000"}/`;
